requirements.txt
fastapi
gunicorn
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-sdk
